+++ src/core/iterator.h
@@ -20,7 +20,7 @@ namespace original {
             void operator--(int) const;
             virtual void operator+=(int64_t steps) const = 0;
             virtual void operator-=(int64_t steps) const = 0;
-            int compareTo(const iterator &other) const override;
+            int64_t compareTo(const iterator &other) const override;
             virtual int64_t operator-(const iterator& other) const = 0;
             iterator* clone() const override = 0;
             explicit operator bool() const;
@@ -96,7 +96,7 @@ namespace original {
     }
 
     template<typename TYPE>
-    auto original::iterator<TYPE>::compareTo(const iterator &other) const -> int {
+    auto original::iterator<TYPE>::compareTo(const iterator &other) const -> int64_t {
         return this->operator-(other);
     }
